Cleaning up the existing dca files
With this commit i am cleaning up the tl_timeline.php
and the config.php. Now everything is indendet with tabs
and it's a little bit better formated in the backend.

Also i changed the language callback to the contao default.
This means that the languages are always displayed in the corrent
backend language.

Signed-off-by: Leo Unglaub <leo.unglaub@2create.at>